requirements.txt
flask
pandas
numpy
scikit-learn
xgboost
statsmodels
pulp
shap
sqlalchemy
flask-cors
pyjwt
